% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_moments.R
\name{bound_moments}
\alias{bound_moments}
\title{Check Fréchet bounds}
\usage{
bound_moments(moments, nu = 1)
}
\arguments{
\item{moments}{\code{matrix} a m x m square matrix, it is assumed that the diagonal gives
probabilies for events A_1,...,A_m and the off diagonal gives}

\item{nu}{\code{numeric}}
}
\value{
A list giving detailed results.
}
\description{
Check Fréchet bounds
}
\details{
See \url{https://en.wikipedia.org/wiki/Fréchet_inequalities}.
}
