\name{bumpers}
\alias{bumpers}
\non_function{}
\title{ bumper repair costs }
\description{ bumper repair costs }
\usage{data(bumpers)}
\format{
  The format is:
 Named num [1:23]  618  795 1304 1308 1340 ...
 - attr(*, "names")= chr [1:23] "Honda Accord" "Chevrolet Cavalier" "Toyota Camry" "Saturn SL2" ...
}
\source{
  From Kitchens Exploring Statistics
}
\examples{
data(bumpers)
stem(bumpers)
}
\keyword{datasets}
