\name{cabinets}
\alias{cabinets}
\non_function{}
\title{ cost of cabinets from 2 suppliers }
\usage{data(cabinets)}
\description{
The \code{cabinets} data frame has 20 rows and 2 columns.

}
\format{
  This data frame contains the following columns:
  \describe{
    \item{SupplierA}{a numeric vector}
    \item{SupplierB}{a numeric vector}
  }
}
\source{
  Borrowed from Kitchen's Exploring Statistics
}
\examples{
data(cabinets)
boxplot(cabinets)
}
\keyword{datasets}
