% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_dist.R
\name{combine_groups}
\alias{combine_groups}
\title{Combine subgroup populations into single \code{SIMPle.dist} object}
\usage{
combine_groups(..., dists = NULL)
}
\arguments{
\item{...}{multiple (named) \code{SIMPle.dist} objects with same number of variables}

\item{dists}{(named) list of further \code{SIMPle.dist} objects}
}
\value{
A \code{SIMPle.dist} object
}
\description{
Combine subgroup populations into single \code{SIMPle.dist} object
}
