% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_gamma.R
\name{construct_gamma}
\alias{construct_gamma}
\title{Construct the gamma parameter for the multivariate Beta (mBeta) distribution.}
\usage{
construct_gamma(dims = 3, nu, moments = NULL, tol = 0.01,
  method = 1, ...)
}
\arguments{
\item{dims}{\code{integer}, dimension of multivariate beta distribution.}

\item{nu}{\code{numeric}, equal to the sum of resulting gamma vector (prior observations).}

\item{moments}{\code{matrix}, .... The entry (j,k) of \code{cp} specifies the
expected joint probability P(B_j = B_k = 1). See details.}

\item{tol}{\code{numeric}, if obtained momemnt matrix deviates more than \code{tol} from the target
a warning is given. See details.}

\item{method}{\code{integer}, ... either 1 (\code{lsei}) or 2 (\code{solve.QP}). Is passed to limSolve::lsei.}
}
\value{
A numeric vector gamma of length 2^dims suitable as a parameter for
the multivariate Beta (mBeta) distribution.
}
\description{
Construct the gamma parameter for the multivariate Beta (mBeta) distribution.
}
\details{
[+++TODO+++] Frechet bounds! incl REF
CP VS MU/RHO SPEC
SIMPle REF
}
\examples{
construct_gamma(3, 1)

}
