% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sample.R
\name{convert_sample}
\alias{convert_sample}
\title{Apply functions to SIMPle.sample objects}
\usage{
convert_sample(sample, margin, fun, args = list())
}
\arguments{
\item{sample}{\code{SIMPle.sample} object, see \code{\link{draw_sample}}}

\item{margin}{either integer vector with values 1,2 or 3 or character vector with values
"obs", "var" or "group" of maximum length 3. Function fun is applied in the specified order.}

\item{fun}{function to be applied}

\item{args}{list of arguments to be passed to \code{fun}}
}
\description{
Apply functions to SIMPle.sample objects
}
