% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2cov.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert a correlation matrix and standard deviations into a covariance matrix.}
\usage{
cor2cov(rho, sigma)
}
\arguments{
\item{rho}{\code{matrix}, the underlying correlation matrix.}

\item{sigma}{\code{numeric}, vector of standard deviations.}
}
\value{
The corresponding covariance matrix.
}
\description{
Convert a correlation matrix and standard deviations into a covariance matrix.
}
\details{
Copied implementation from \code{psych::cor2cov}, adapted to work with length(sigma)==1
}
