% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_Beta.R
\name{define_Beta}
\alias{define_Beta}
\title{Define a univariate Beta distribution.}
\usage{
define_Beta(alpha = 1, beta = 1, vars = NULL)
}
\arguments{
\item{alpha}{\code{numeric}, first shape parameter \eqn{\alpha \ge 0}.}

\item{beta}{\code{numeric}, second shape parameter \eqn{\beta \ge 0}.}
}
\value{
A \code{SIMPle.dist} object representing the univariate Beta
distribution with given parameters alpha and beta.
}
\description{
Define a \code{SIMPle} representation of a univariate Beta distribution via its two parameters.
}
\details{
A extensive overview over the univariate Beta distribution is given at
\url{https://en.wikipedia.org/wiki/Beta_distribution}.
}
\examples{
define_Beta(1/2, 1/2) # "Jeffrey's prior distribution"

}
