% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dist.R
\name{define_dist}
\alias{define_dist}
\title{Define a multivariate beta (mbeta) distribution}
\usage{
define_dist(
  nu = 1,
  mean = 1/2,
  corr = 0,
  moments = NULL,
  gamma = NULL,
  vars = length(mean),
  varnames = NULL,
  groups = 1,
  groupnames = NULL,
  mode = c("auto", "full", "reduced"),
  tol = 0.01,
  project = TRUE,
  msg = TRUE
)
}
\arguments{
\item{nu}{\code{integer}, 'pseudo-prior-observations'}

\item{mean}{\code{numeric} of length vars, specifying the marginal means. Repeated if of length one.}

\item{corr}{\code{matrix}, symetric, positive definite correlation matrix. If a single correlation}

\item{moments}{\code{matrix}, symmetric matrix with mixed 2nd order moments of the target distribution. See details.}

\item{gamma}{numeric vector, parameter of the underlying Dirichlet distribution.
the base 2 logarithm of length(gamma) needs to be an integer and is the dimension of the
resulting multivariate Beta distribution.}

\item{vars}{\code{integer}, number of variables}

\item{varnames}{\code{character}, variables names, needs to have length \code{vars}}

\item{groups}{\code{integer}, number of groups (subpopulations) (default: \code{1})}

\item{groupnames}{\code{character}, group names}

\item{mode}{\code{character}, either \code{"auto"} (default), \code{"full"} or \code{"reduced"}. See details.}

\item{tol}{\code{numeric}, maximum relative deviation (default: 0.01) between specified and
achieved moment matrix (or correlation matrix)}

\item{project}{\code{logical}, project parameters to satisfy Frechet bounds
(\code{TRUE}, default) or not (\code{FALSE})}

\item{msg}{\code{logical}, return messages (\code{TRUE}, default) or not (\code{FALSE})}
}
\value{
A SIMPle.dist object representing the multivariate Beta
distribution with given parameters.
}
\description{
Define a multivariate beta (mbeta) distribution
}
\details{
Mode "full" implies that the full parameter vector gamma is used which results in nonzero
mixed moments. In mode "reduced", a sparse representation with 1+m+(m-1)*m/2 parameters is used to
model (prior) sample size, mean and covariance / correlation matrix. Per default (mode="auto"),
the mode is set to "full" for dimension <= 10 and "reduced" for dimensions >10.
}
\examples{
define_dist(vars = 3, nu=10, mean=c(0.9,0.8,0.7), corr=0.5)
}
