% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dlist.R
\name{define_dlist}
\alias{define_dlist}
\title{Define a list ob SIMPle.dist objects.}
\usage{
define_dlist(..., dlist = NULL)
}
\arguments{
\item{...}{\code{SIMPle.dist} objects}

\item{dlist}{\code{list}, a list of \code{SIMPle.dist} objects}
}
\value{
A \code{SIMPle.dlist}, a list ob \code{SIMPle.dist} obejcts.
}
\description{
Specify different distributions, e.g. related to the same varibale(s)
on different subsets of the data.
}
\examples{
# Define priors for sensitivity (se) and specificity (sp) of a binary classifier:
define_dlist(se = define_Beta(4,1), sp = define_Beta(3,2))
}
