% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_loss.R
\name{define_loss}
\alias{define_loss}
\title{Define a loss}
\usage{
define_loss(name = "fwer", loss = fp, cpe = sum, aggr = max,
  control = list())
}
\arguments{
\item{name}{\code{character}, name of the loss. All other arguments are ignored if the name
for a known loss}

\item{loss}{\code{function}, needs to have formal arguments [+++ TODO +++].}

\item{cpe}{\code{function}, specifies if co-primary enpoints ...}

\item{control}{\code{list}, can be used to specify default arguments of lossfun.}

\item{max}{\code{function}, specifies how per variable loss is aggregated over variables}
}
\value{
A \code{SIMPle.loss} object, may be used for statistical \code{\link{inference}}.
}
\description{
Define a loss
}
\examples{
[+++ TODO +++]
}
