% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_mBeta.R
\name{define_mBeta}
\alias{define_mBeta}
\title{Define a multivariate Beta (mBeta) distribution}
\usage{
define_mBeta(dims = 2, nu = 1, mean = 1/2, corr = 0,
  moments = NULL, gamma = NULL, vars = NULL, mode = c("auto",
  "full", "reduced"), tol = 0.01, project = TRUE, msg = TRUE)
}
\arguments{
\item{dims}{\code{integer}, number of dimensions}

\item{nu}{\code{integer}, 'pseudo-prior-observations'}

\item{mean}{\code{numeric} of length dims, specifying the marginal means. Repeated if of length one.}

\item{corr}{\code{matrix}, symetric, positive definite correlation matrix. If a single correlation}

\item{moments}{\code{matrix}, symmetric matrix with mixed 2nd order moments of the target distribution. See details.}

\item{gamma}{numeric vector, parameter of the underlying Dirichlet distribution.
the base 2 logarithm of length(gamma) needs to be an integer and is the dimension of the
resulting multivariate Beta distribution.}

\item{vars}{\code{character}, of length \code{dims}}

\item{mode}{\code{character}, either \code{"auto"} (default), \code{"full"} or \code{"reduced"}. See details.}

\item{tol}{\code{numeric}, maximum relative deviation (default: 0.01) between specified and
achieved moment matrix (or correlation matrix)}
}
\value{
A SIMPle.dist object representing the multivariate Beta
distribution with given parameters.
}
\description{
Define a multivariate Beta (mBeta) distribution
}
\details{
Mode "full" implies that the full parameter vector gamma is used which results in nonzero
mixed moments. In mode "reduced", a sparse representation with 1+m+(m-1)*m/2 parameters is used to
model (prior) sample size, mean and covariance / correlation matrix. Per default (mode="auto"),
the mode is set to "full" for dimension <= 10 and "reduced" for dimensions >10.
}
\examples{
define_mBeta(dims = 3, nu=10, mean=c(0.9,0.8,0.7), corr=0.5)
}
