% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample.R
\name{draw_sample}
\alias{draw_sample}
\title{Draw a sample from a (multivariate) probability distribution}
\usage{
draw_sample(dist, n = 10000, method = "auto", msg = TRUE, ...)
}
\arguments{
\item{dist}{SIMPle.dist or SIMPle.dlist, specifies distribution to sample from}

\item{n}{integer; number of sampled observations}

\item{method}{character; specifies sampling method. Either "mcmc" for Monte Carlo (MC)
Simulation or "copula" for univariate MC + Copula approach}
}
\value{
A data matrix representing the sample
}
\description{
Draw a sample from a (multivariate) probability distribution
}
