% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample_pred.R
\name{draw_sample_pred}
\alias{draw_sample_pred}
\title{Draw a sample from the predictive (mbeta-mbin) distribution}
\usage{
draw_sample_pred(
  dist,
  sample = 100,
  size = rep(1, groups(dist)),
  method_pred = c("mbeta_approx", "rmvbin"),
  method_ext = c("basic", "cov"),
  count = TRUE,
  msg = TRUE
)
}
\arguments{
\item{dist}{\code{SIMPle.dist} object, representing a multivariate beta distribution see \code{\link{define_dist}}}

\item{sample}{either a precomputed \code{SIMPle.sample} object or an integer specifying the sample size}

\item{size}{sample size(s) for predictive count data}

\item{method_pred}{character, for drawing predictive sample, either "mbeta_approx" (recommended) or "rmvbin"}

\item{method_ext}{character, method for sample 'extension', either "basic" (default) or "cov"}

\item{count}{logical, should outcome be counts (TRUE; default) or probabilities (FALSE)}

\item{msg}{logical, should messages be displayed? (default: TRUE)}
}
\description{
Draw a sample from the predictive (mbeta-mbin) distribution
}
\details{
Approximation of predictive mbeta-mbin (multivariate beta - multivariate binomial) distribution
}
