\name{exec.pay}
\alias{exec.pay}
\non_function{}
\title{ Direct compensation for 200 United States CEO's for 2000  }
\description{ Direct compensation for 200 United States CEO's for 2000  }
\usage{data(exec.pay)}
\format{
  The format is:
 num [1:199] 136 74 8 38 46 43 9 9 12 11 ...
}
\source{
New York Times Business section 04/01/2001
}
\examples{
data(exec.pay)
hist(exec.pay)
}
\keyword{datasets}
