% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnGSQDesign.R
\name{fnGSQDesign}
\alias{fnGSQDesign}
\title{Creates a simple platform trial design using a GSQ plan and continuous endpoints of class lPltfDsgn}
\usage{
fnGSQDesign(
  dMuC = 0,
  dSigC = 1,
  vMuT = 1,
  vSigT = 1,
  nOutObsLag = 0,
  nRecrSpeed = 10,
  nMaxISA = 10,
  nMaxNIntr = 20,
  vAnalysisTimes = c(1),
  nIntrInitial = 1,
  bReplaceIntr = FALSE,
  dProbReplaceIntr = 1,
  nWeeksNewIntr = NA,
  dProbWeeksNewIntr = 1,
  cAllocRatio = "balanced",
  vEffBounds = c(0.05),
  vFutBounds = c(0.05)
)
}
\arguments{
\item{dMuC}{Mean in the control group; by default 0}

\item{dSigC}{Standard Deviation in the control group; by default 1}

\item{vMuT}{Vector of means in the treatment groups; if not equal in length to number of
investigated treatments, vector will be recycled; be default 1}

\item{vSigT}{Vector of stardard deviations in the treatment groups; if not equal in length
to number of investigated treatments, vector will be recycled; be default 1}

\item{nOutObsLag}{Time lag in observing patient outcomes; by default 0}

\item{nRecrSpeed}{How many patients per time unit are enrolled? Default is 10}

\item{nMaxISA}{Maximum number of ISAs to include in platform; Default is 10}

\item{nMaxNIntr}{Maximum sample size in treatment arms; Default is 20}

\item{vAnalysisTimes}{Vector giving timings of interim analyses in information fractions; Default is c(1)}

\item{nIntrInitial}{Number of treatments that start in the platform trial alongside control; Default 1}

\item{bReplaceIntr}{Should outgoing treatments be replaced with new treatments? Default is FALSE}

\item{dProbReplaceIntr}{Probability with which outgoing treatments should be replaced. Default is 1.}

\item{nWeeksNewIntr}{After how many weeks should Treatments be included?
If NULL, treatments are not included based on time. Default is NULL.}

\item{dProbWeeksNewIntr}{Probability with which treatments are included based on time. Default is 1.}

\item{cAllocRatio}{How are patients allocated to treatments. Options are "balanced" or "squareroot".
Default is balanced.}

\item{vEffBounds}{Vector of same length as vAnalysisTimes that gives the efficacy bounds at the
different analysis time points}

\item{vFutBounds}{Vector of same length as vAnalysisTimes that gives the futility bounds at the
different analysis time points}
}
\value{
Object of class lPltfDsgn to be supplied to fnRunSingleTrialSim or modified later
}
\description{
Creates a simple platform trial design using a GSQ plan and continuous endpoints of class lPltfDsgn
}
\examples{

lPltfDsgn <- 
 fnGSQDesign( 
  dMuC              = 0,
  dSigC             = 1,
  vMuT              = 0.25,
  vSigT             = 1,
  nOutObsLag        = 2,
  nRecrSpeed        = 5,
  nMaxISA           = 4,
  nMaxNIntr         = 50,
  vAnalysisTimes    = c(0.5, 1),
  nIntrInitial      = 1,
  bReplaceIntr      = TRUE,
  dProbReplaceIntr  = 0.5,
  nWeeksNewIntr     = 2,
  dProbWeeksNewIntr = 0.5,
  cAllocRatio       = "balanced",
  vEffBounds        = c(0.01, 0.04),
  vFutBounds        = c(0.25, 0.05)
 )

out1 <- fnRunSingleTrialSim(lPltfDsgn, bRetainSnaps = TRUE)
ocs1 <- fnSimDsgnOC(lPltfDsgn = lPltfDsgn, nIter = 5)
ocs2 <- fnSimDsgnOC(lPltfDsgn = lPltfDsgn, nIter = 300, nCores = 7)
out2 <- fnRunSingleTrialSim(fnGSQDesign())

}
