% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRunSingleTrialSim.R
\name{fnRunSingleTrialSim}
\alias{fnRunSingleTrialSim}
\title{Runs a single platform trial simulation}
\usage{
fnRunSingleTrialSim(
  lPltfDsgn,
  bRetainSnaps = TRUE,
  bCreateLog = FALSE,
  cLogName = "platform"
)
}
\arguments{
\item{lPltfDsgn}{Platform Design List}

\item{bRetainSnaps}{Whether or not to keep the list of snapshots; Default is TRUE}

\item{bCreateLog}{Whether or not to create a log file in current folder; Default is TRUE}

\item{cLogName}{Name of Log File; Default is "platform"}
}
\description{
Runs a single platform trial simulation
}
