% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnSimDsgnOC.R
\name{fnSimDsgnOC}
\alias{fnSimDsgnOC}
\title{Calculates the operating characteristics of a platfom design}
\usage{
fnSimDsgnOC(
  lPltfDsgn,
  nIter,
  nCores = 1,
  bSaveCsv = FALSE,
  cPathCsv = getwd(),
  cNameCsv = "Test",
  bRetList = FALSE,
  bRetInd = FALSE,
  export = NULL,
  nStartSeed = NULL,
  ...
)
}
\arguments{
\item{lPltfDsgn}{Platform Design List}

\item{nIter}{Number of iterations to run}

\item{nCores}{Number of cores to run on}

\item{bSaveCsv}{Indicator whether simulation results should be saved in a CSV file}

\item{cPathCsv}{Path to which simulation results will be saved; if NULL, then save to current path}

\item{cNameCsv}{Filename for CSV File to be saved as}

\item{bRetList}{Indicator whether function should save list of results}

\item{bRetInd}{Indicator whether individual trial results should be saved as well}

\item{export}{Further Arguments to be exported to slaves}

\item{nStartSeed}{Only used when nCores == 1, used for debugging}

\item{...}{All other design parameters}
}
\value{
List of operating characteristics
}
\description{
Calculates the operating characteristics of a platfom design
}
\examples{

ocs <- fnSimDsgnOC(lPltfDsgn = fnSimpleDesign(), nIter = 3)

}
