% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTestDesign.R
\name{fnTestDesign}
\alias{fnTestDesign}
\title{Creates a simple platform trial design of class lPltfDsgn}
\usage{
fnTestDesign(
  effect_scenario = 1,
  recr_speed = 20/4.25,
  ISA_add = 12,
  max_time = 255,
  futility_stop = TRUE,
  ISA_start = 3,
  lag_units = 6,
  alpha = 0.05
)
}
\arguments{
\item{effect_scenario}{One of the three effect scenarios specified (1 == all not effective,
2 == all moderately effective, 3 == Rotation between effects)}

\item{recr_speed}{How many patients per time unit are enrolled? Poisson distribution}

\item{ISA_add}{After how many weeks will a new treatment enter}

\item{max_time}{Number of weeks until which ISAs can be added to platform trial}

\item{futility_stop}{Boolean; whether or not the trial can be stopped at interim for futility}

\item{ISA_start}{Number of ISAs when trial is initiated}

\item{lag_units}{Number of time units until outcome is observed}

\item{alpha}{Significance level for one-sided p-value}
}
\value{
Object of class lPltfDsgn to be supplied to fnRunSingleTrialSim or modified later
}
\description{
Creates a simple platform trial design of class lPltfDsgn
}
\examples{

lPltfDsgn <- 
 fnTestDesign(
   effect_scenario = 3,
   recr_speed = 28,
   ISA_add = 2,
   max_time = 0,
   futility_stop = TRUE,
   ISA_start = 3,
   lag_units = 1
 )
 
ocs <- fnSimDsgnOC(lPltfDsgn = lPltfDsgn, nIter = 300, nCores = 7)

out1 <- fnRunSingleTrialSim(lPltfDsgn, bRetainSnaps = TRUE)
ocs1 <- fnSimDsgnOC(lPltfDsgn = lPltfDsgn, nIter = 5)
out2 <- fnRunSingleTrialSim(fnTestDesign())

}
