% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnWP6NCC.R
\name{fnWP6NCC}
\alias{fnWP6NCC}
\title{Creates a simple platform trial design of class lPltfDsgn}
\usage{
fnWP6NCC(
  n_fin = 150,
  cohorts_start = 2,
  rr_comb = 0.35,
  rr_plac = 0.1,
  cohort_fixed = 24,
  cohorts_max = 5,
  sharing_type = "All",
  accrual_param = 6,
  time_trend = 0,
  Bayes_Sup = matrix(c(0, 0.95, 0.05, 0.8, 0.1, 0.6), nrow = 3, ncol = 2, byrow = TRUE)
)
}
\arguments{
\item{n_fin}{Sample size per cohort at final}

\item{cohorts_start}{Number of cohorts to start the platform with}

\item{rr_comb}{Response rates of treatment}

\item{rr_plac}{Response rate of the SoC}

\item{cohort_fixed}{If not NULL, fixed timesteps after which a cohort will be included}

\item{cohorts_max}{Maximum number of cohorts that are allowed to be added throughout the trial}

\item{sharing_type}{Which control data should be shared. Default is "All", meaning all available data will be used for analysis. Other options include "NCC", which will use NCC data for evidence levels 2&3. Finally, "CC" will use only concurrent control data for all levels of evidence.}

\item{accrual_param}{Number of patients per week}

\item{time_trend}{Change in response rate over time}

\item{Bayes_Sup}{Efficacy Decision Rules}

\item{...}{Further arguments to be passed to decision function, such as decision making criteria}
}
\value{
Object of class lPltfDsgn to be supplied to fnRunSingleTrialSim or modified later
}
\description{
Creates a simple platform trial design of class lPltfDsgn
}
\examples{

# Comparison IA1
Bayes_Sup <- matrix(nrow = 3, ncol = 2)
Bayes_Sup[1,] <- c(0.00, 0.95)
Bayes_Sup[2,] <- c(0.10, 0.80)
Bayes_Sup[3,] <- c(0.20, 0.60)

lPltfDsgn <- fnWP6NCC(Bayes_Sup = Bayes_Sup)


Bayes_Sup <- matrix(nrow = 1, ncol = 2)
Bayes_Sup[1,] <- c(0.00, 0.95)
lPltfDsgn <- fnWP6NCC(Bayes_Sup = Bayes_Sup)

out <- fnRunSingleTrialSim(lPltfDsgn)
ocs1 <- fnSimDsgnOC(lPltfDsgn = lPltfDsgn, nIter = 20)

}
