% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnWP6simple.R
\name{fnWP6simple}
\alias{fnWP6simple}
\title{Creates a simple platform trial design of class lPltfDsgn}
\usage{
fnWP6simple(
  n_fin = 150,
  cohorts_start = 2,
  rr_comb1 = 0.35,
  rr_comb2 = 0.35,
  rr_plac1 = 0.1,
  rr_plac2 = 0.2,
  correlation = 0,
  cohort_fixed = 24,
  cohort_random = NULL,
  cohorts_max = 5,
  sharing_type = "Conc",
  accrual_param = 6,
  analysis_times = c(0.5, 0.75, 1),
  hist_lag = 52,
  time_trend = 0,
  composite = "or",
  Bayes_Sup1 = list(matrix(c(0, 0.95), nrow = 1), matrix(c(0, 0.95), nrow = 1),
    matrix(c(0, 0.95), nrow = 1)),
  Bayes_Sup2 = list(matrix(c(0, 0.95), nrow = 1), matrix(c(0, 0.95), nrow = 1),
    matrix(c(0, 0.95), nrow = 1)),
  Bayes_Fut1 = NULL,
  Bayes_Fut2 = NULL
)
}
\arguments{
\item{n_fin}{Sample size per cohort at final}

\item{cohorts_start}{Number of cohorts to start the platform with}

\item{rr_comb1}{Response rates of treatment, histology endpoint 1}

\item{rr_comb2}{Response rates of treatment, histology endpoint 2}

\item{rr_plac1}{Response rate of the SoC, histology endpoint 1}

\item{rr_plac2}{Response rate of the SoC, histology endpoint 2}

\item{correlation}{Correlation between histology endpoints}

\item{cohort_fixed}{If not NULL, fixed timesteps after which a cohort will be included}

\item{cohort_random}{If not NULL, probability of including new cohorts at every time unit. If this is
specified, cohort_fixed is ignored.}

\item{cohorts_max}{Maximum number of cohorts that are allowed to be added throughout the trial}

\item{sharing_type}{Which backbone and placebo data should be used for arm comparisons; Default is "All".
Another option is "Conc" or "Intr".}

\item{accrual_param}{Number of patients per week}

\item{analysis_times}{Vector of information fractions needed for interim analyses}

\item{hist_lag}{Time until histology outcome is observed}

\item{time_trend}{Additive term by which response rates increase at every time step}

\item{composite}{Rule for deriving the composite endpoint. By default "or", otherwise "and"}

\item{Bayes_Sup1}{Efficacy Rules for Endpoint 1}

\item{Bayes_Sup2}{Efficacy Rules for Endpoint 2}

\item{Bayes_Fut1}{Futility Rules for Endpoint 1}

\item{Bayes_Fut2}{Futility Rules for Endpoint 2}

\item{...}{Further arguments to be passed to decision function, such as decision making criteria}
}
\value{
Object of class lPltfDsgn to be supplied to fnRunSingleTrialSim or modified later
}
\description{
Creates a simple platform trial design of class lPltfDsgn
}
\examples{

# Comparison IA1
Bayes_Sup11 <- matrix(nrow = 2, ncol = 2)
Bayes_Sup11[1,] <- c(0.00, 0.95)
Bayes_Sup11[2,] <- c(0.10, 0.80)
# Comparison IA2
Bayes_Sup12 <- matrix(nrow = 2, ncol = 2)
Bayes_Sup12[1,] <- c(0.00, 0.95)
Bayes_Sup12[2,] <- c(0.10, 0.80)
# Comparison IA3
Bayes_Sup13 <- matrix(nrow = 2, ncol = 2)
Bayes_Sup13[1,] <- c(0.00, 0.95)
Bayes_Sup13[2,] <- c(0.10, 0.80)

Bayes_Sup1 <- Bayes_Sup2 <- list(Bayes_Sup11, Bayes_Sup12, Bayes_Sup13)

# Comparison IA1
Bayes_Fut11 <- matrix(nrow = 1, ncol = 2)
Bayes_Fut11[1,] <- c(0.00, 0.20)
# Comparison IA2
Bayes_Fut12 <- matrix(nrow = 1, ncol = 2)
Bayes_Fut12[1,] <- c(0.00, 0.30)
# Comparison IA3
Bayes_Fut13 <- NULL
# Endpoint 1+2
Bayes_Fut1 <- Bayes_Fut2 <- list(Bayes_Fut11, Bayes_Fut12, Bayes_Fut13)

lPltfDsgn <- fnWP6simple()

out <- fnRunSingleTrialSim(lPltfDsgn)
ocs1 <- fnSimDsgnOC(lPltfDsgn = lPltfDsgn, nIter = 5)

}
