% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get}
\alias{get}
\title{Extract information}
\usage{
get(dist, attr = "type")
}
\arguments{
\item{dist}{\code{SIMPle.dist}}

\item{attr}{\code{character}}
}
\value{
[+++TODO+++]
}
\description{
Extract information
}
\examples{
[+++TODO+++]
}
