% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Construct a credible region for a (multivariate) probability distribution}
\usage{
inference(dist, loss = "fwer", target = 0.05,
  alternative = c("two.sided", "greater", "less"), transfrom = NULL,
  threshold = 0.8, control = list(), method = c("auto", "approx",
  "copula", "mcmc"), type = "equi.tail", n.rep = 10000, seed = 1,
  ...)
}
\arguments{
\item{dist}{\code{SIMPle.dist} object.}

\item{loss}{\code{character}, known loss (e.g. "fwer") or self constructed loss.
(see \code{\link{define_loss}}).}

\item{target}{\code{numeric}, target value for loss, e.g. one minus coverage probability for
\code{loss="fwer"}.}

\item{alternative}{\code{character}.}

\item{transfrom}{[currently unused]}

\item{threshold}{[currently unused]}

\item{control}{\code{list}, contains further arguments for loss.}

\item{method}{\code{character}, specfies the method used, see details.}

\item{type}{\code{character}, type of credible region, currently only \code{"equi.tail"} is supported.}

\item{n.rep}{\code{integer}, number of replications to draw from \code{dist} for mcmc methods}

\item{seed}{\code{integer}, random seed to make result replicable.}

\item{...}{}
}
\value{
A \code{SIMPle.dist} object with entry \code{inference}. Call \code{summary} or
\code{visualize} on output to investiagte result.
}
\description{
Construct a credible region for a (multivariate) probability distribution
}
\details{
[+++TODO+++]
}
\examples{
[+++TODO+++]
}
