% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lAddIntr.R
\name{lAddIntr}
\alias{lAddIntr}
\alias{new_lAddIntr}
\alias{validate_lAddIntr}
\alias{summary.lAddIntr}
\title{Addition of ISAs into ongoing platform trial}
\usage{
new_lAddIntr(
  fnAddIntr = function(lPltfDsgn, lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lAddIntr(x)

lAddIntr()

\method{summary}{lAddIntr}(object, ...)
}
\arguments{
\item{fnAddIntr}{Function which will add ISAs}

\item{lAddArgs}{Further arguments used in fnAddIntr}

\item{x}{An object of class lAddIntr}

\item{object}{An object of class lAddIntr}

\item{...}{Additional arguments (not used)}
}
\description{
Functions for creating, validating and simple use of class lAddIntr
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lAddIntr()
}

}
