% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lAddPats.R
\name{lAddPats}
\alias{lAddPats}
\alias{new_lAddPats}
\alias{validate_lAddPats}
\alias{summary.lAddPats}
\title{Add simulated Pats to ISAs}
\usage{
new_lAddPats(
  fnAddPats = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lAddPats(x)

lAddPats()

\method{summary}{lAddPats}(object, ...)
}
\arguments{
\item{fnAddPats}{Function that handles adding patients to respective ISAs}

\item{lAddArgs}{Further arguments used in fnAddPats}

\item{x}{An object of class lAddPats}

\item{object}{An object of class lAddPats}

\item{...}{Additional arguments (not used)}
}
\description{
Functions and rules for adding Pats to ISAs of class lAddPats
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lAddPats()
}

}
