% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lAllocIntr.R
\name{lAllocIntr}
\alias{lAllocIntr}
\alias{new_lAllocIntr}
\alias{validate_lAllocIntr}
\alias{summary.lAllocIntr}
\title{Between ISA allocation ratio}
\usage{
new_lAllocIntr(
  fnAllocIntr = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lAllocIntr(x)

lAllocIntr()

\method{summary}{lAllocIntr}(x, ...)
}
\arguments{
\item{fnAllocIntr}{Function that handles allocation of patients between ISAs}

\item{lAddArgs}{Further arguments used in fnAllocIntr}
}
\description{
Functions and rules for between ISA allocation ratio of class lAllocIntr
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lAllocIntr()
}

}
