% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lAnls.R
\name{lAnls}
\alias{lAnls}
\alias{new_lAnls}
\alias{validate_lAnls}
\title{Run Analysis}
\usage{
new_lAnls(
  fnAnls = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lAnls(x)

lAnls(
  endpoint = "binary",
  analysis_function_binary = function(x) {
     matOutcome <- table(x$Arm, x$Outcome)
   
     if (ncol(matOutcome) == 1) {
         ret <- 1
     }
     else (ret <-
    stats::prop.test(matOutcome)$p.value)
     return(ret)
 },
  analysis_function_continuous = function(x) {
     stats::t.test(Outcome ~ Arm, data =
    x)$p.value
 },
  group1 = c("C", "All"),
  group2
)
}
\arguments{
\item{fnAnls}{Function that handles running statistical analyses}

\item{lAddArgs}{Further arguments used in fnAnls}

\item{endpoint}{Type of endpoint ("binary" or "continuous")}

\item{analysis_function_binary}{Analysis function for binary endpoints}

\item{analysis_function_continuous}{Analysis function for continuous endpoints}

\item{group1}{First group for comparison}

\item{group2}{Second group for comparison}
}
\description{
Functions and rules for conducting analyses of class lAnls
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lAnls(endpoint = "binary")
}

}
