% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{launch_application}
\alias{launch_application}
\title{Launch \code{reverse_string} shiny application.}
\usage{
launch_application(...)
}
\arguments{
\item{...}{Arugments to \code{shiny::runApp()}.}
}
\description{
This is a wrapper function for launching the 'application' shiny app in this
package.
}
\examples{
launch_application()
}

