% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lCheckEnrl.R
\name{lCheckEnrl}
\alias{lCheckEnrl}
\alias{new_lCheckEnrl}
\alias{validate_lCheckEnrl}
\alias{summary.lCheckEnrl}
\title{Check ISA Enrollment}
\usage{
new_lCheckEnrl(
  fnCheckEnrl = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lCheckEnrl(x)

lCheckEnrl()

\method{summary}{lCheckEnrl}(x, ...)
}
\arguments{
\item{fnCheckEnrl}{Function that checks if an ISA is still actively enrolling}

\item{lAddArgs}{Further arguments used in fnCheckEnrl}
}
\description{
Functions and rules for within ISA checking of Enrollment of class lCheckEnrl
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lCheckEnrl()
}

}
