% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lOCSynth.R
\name{lOCSynth}
\alias{lOCSynth}
\alias{new_lOCSynth}
\alias{validate_lOCSynth}
\title{Synthesize individual trial simulations into OCs}
\usage{
new_lOCSynth(
  fnOCSynth = function(lIndTrials, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lOCSynth(x)

lOCSynth()
}
\arguments{
\item{fnOCSynth}{Function that synthesizes individual trial results into operating characteristics}

\item{lAddArgs}{Further arguments used in fnOCSynth}
}
\description{
Functions and rules for synthesizing results of class lOCSynth
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lOCSynth()
}

}
