% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lPltfSummary.R
\name{lPltfSummary}
\alias{lPltfSummary}
\alias{new_lPltfSummary}
\alias{validate_lPltfSummary}
\title{Summarize the platform trial}
\usage{
new_lPltfSummary(
  fnPltfSummary = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lPltfSummary(x)

lPltfSummary()
}
\arguments{
\item{fnPltfSummary}{Function which will check if Stop Rules are reached}

\item{lAddArgs}{Further arguments used in fnPltfSummary}

\item{x}{An object of class lPltfSummary}
}
\description{
Functions for summarizing the platform trial of class lPltfSummary
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lPltfSummary()
}

}
