% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lRecrPars.R
\name{lRecrPars}
\alias{lRecrPars}
\alias{new_lRecrPars}
\alias{validate_lRecrPars}
\alias{plot.lRecrPars}
\alias{summary.lRecrPars}
\title{Recruitment Parameters}
\usage{
new_lRecrPars(
  fnRecrProc = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lRecrPars(x)

lRecrPars(nPat)

\method{plot}{lRecrPars}(x, dCurrTime = 1:52, ...)

\method{summary}{lRecrPars}(object, ...)
}
\arguments{
\item{fnRecrProc}{Function which simulates number of new patient arrivals}

\item{lAddArgs}{Further arguments used in recruitment function}

\item{x}{An object of class lRecrPars}

\item{nPat}{Number of patients}

\item{dCurrTime}{Vector of current time steps (default 1:52)}

\item{...}{Additional arguments (not used)}

\item{object}{An object of class lRecrPars}
}
\description{
Functions for creating, validating and simple use of class lRecrPars
}
\examples{

x <- lRecrPars(4)
validate_lRecrPars(x)
plot(x)
summary(x)

}
