% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lSimBase.R
\name{lSimBase}
\alias{lSimBase}
\alias{new_lSimBase}
\alias{validate_lSimBase}
\title{Simulation of patient Baseline data}
\usage{
new_lSimBase(
  fnSimBase = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lSimBase(x)

lSimBase(vars = list(list("time", "t")), names = c("InclusionTime"))
}
\arguments{
\item{fnSimBase}{Function that simulates patient baseline characteristics}

\item{lAddArgs}{Further arguments used in fnSimBase}

\item{x}{An object of class lSimBase}

\item{vars}{List of variables to simulate}

\item{names}{Names for the simulated variables}
}
\description{
Functions and rules for simulating patient Baseline data of class lSimBase
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lSimBase()
}

}
