% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lUpdIntrAlloc.R
\name{lUpdIntrAlloc}
\alias{lUpdIntrAlloc}
\alias{new_lUpdIntrAlloc}
\alias{validate_lUpdIntrAlloc}
\title{Update between ISA allocation ratio}
\usage{
new_lUpdIntrAlloc(
  fnUpdIntrAlloc = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lUpdIntrAlloc(x)

lUpdIntrAlloc()
}
\arguments{
\item{fnUpdIntrAlloc}{Function that updates allocation ratios between ISAs}

\item{lAddArgs}{Further arguments used in fnUpdIntrAlloc}

\item{x}{An object of class lUpdIntrAlloc}
}
\description{
Functions and rules for between ISA allocation ratio update of class lUpdIntrAlloc
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lUpdIntrAlloc()
}

}
