% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{margin_params}
\alias{margin_params}
\title{Return marginal parameters of distribution}
\usage{
margin_params(dist, groups = 1:length(dist), simplify = TRUE)
}
\arguments{
\item{dist}{a \code{SIMPle.dist} object}

\item{groups}{integer, restrict query to (sub)group}

\item{simplify}{logical, should output be simplified if length(groups)==1 (default: TRUE)}
}
\description{
Return marginal parameters of distribution
}
