\name{math}
\alias{math}
\non_function{}
\title{ Standardized math scores }
\description{ Standardized math scores }
\usage{data(math)}
\format{
  The format is:
 num [1:30] 44 49 62 45 51 59 57 55 70 64 ...
}
\source{
  From Kitchens Exploring Statistics
}
\examples{
data(math)
hist(math)
}
\keyword{datasets}
