% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_dist.R
\name{order_dist}
\alias{order_dist}
\title{Order margins of a multivariate distribution}
\usage{
order_dist(
  dist,
  method = c("param", "sample"),
  target = c("mean", "prob", "tstat"),
  combine = ifelse(target == "prob", "prod", "min"),
  threshold = rep(0.75, groups(dist)),
  size = NA,
  ties.method = "random"
)
}
\arguments{
\item{dist}{\code{SIMPle.dist} object}

\item{method}{character, base order on dist parameters ("params") or sample ("sample")}

\item{target}{character, basis for ordering}

\item{combine}{character, way to combine across groups}

\item{threshold}{numeric, threshold values (vector of length groups(dist))}

\item{size}{integer, sample size(s) in predictive sample}

\item{ties.method}{character, see argument ties.method from function \code{\link{base::rank}}}
}
\description{
Order margins of a multivariate distribution
}
