% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_dist.R
\name{rank_dist}
\alias{rank_dist}
\title{Ranking a SIMPle.dist (!experimental!)}
\usage{
rank_dist(dist, method = c("prob", "mean"), cpe = c("min", "prod"),
  cu = 0.75, lower.tail = FALSE)
}
\arguments{
\item{dist}{SIMPle.dist object}

\item{method}{"mean" or "prob"}

\item{cpe}{"min" or "prod"}

\item{cu}{numeric, cutoff for method "prob"}

\item{lower.tail}{logical, default FALSE}
}
\value{
ranking
}
\description{
Ranking a SIMPle.dist (!experimental!)
}
