% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverse_string}
\alias{reverse_string}
\title{Reverse a character string.}
\usage{
reverse_string(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
The character string \code{x} reversed.
}
\description{
Reverse a character string.
}
\examples{
reverse_string("hello")
reverse_string("hello world")
}

