% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Draw a sample from a (multivariate) probability distribution}
\usage{
sample(dist, method = "auto", n = "auto", seed = 1, msg = TRUE,
  ...)
}
\arguments{
\item{dist}{SIMPle.Dist or SIMPle.mDist, specifies distribution to sample from}

\item{method}{character; specifies sampling method. Either "mcmc" for Monte Carlo (MC)
Simulation or "copula" for univariate MC + Copula approach}

\item{n}{integer; number of sampled observations}
}
\value{
A data matrix of size Nrep x dist$dim
}
\description{
Draw a sample from a (multivariate) probability distribution
}
\examples{
+++ TODO +++
}
