% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_dist.R
\name{sample_dist_marginal}
\alias{sample_dist_marginal}
\title{Sample from distribution ignoring the dependency structure}
\usage{
sample_dist_marginal(dist, n = 10000)
}
\arguments{
\item{dist}{a SIMPle.dist object}

\item{n}{integer, the sample size}
}
\value{
\code{SIMPle.sample} object
}
\description{
Sample from distribution ignoring the dependency structure
}
