% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dist.R
\name{select_dist}
\alias{select_dist}
\title{Select variables of a distribution (after reordering)}
\usage{
select_dist(dist, method = "mean", cpe = "min", j = NULL,
  m.max = NULL)
}
\arguments{
\item{dist}{SIMPle.dist}

\item{method}{"mean"}

\item{cpe}{"min"}

\item{j}{prespecified ordering}

\item{m.max}{only select first m.max varibales according to ordering}
}
\value{
A SIMPle.dist, potentially of smaller dimension
}
\description{
Select variables of a distribution (after reordering)
}
