% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_groups.R
\name{select_groups}
\alias{select_groups}
\title{Select (sub)groups of a distribution}
\usage{
select_groups(dist, selection)
}
\arguments{
\item{dist}{\code{SIMPle.dist} object}

\item{selection}{integer vector, describes new group order/selection}
}
\value{
A \code{SIMPle.dist} object, potentially with less groups
}
\description{
Select (sub)groups of a distribution
}
