% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vars.R
\name{select_vars}
\alias{select_vars}
\title{Select and reorder variables of a distribution}
\usage{
select_vars(dist, selection = NULL, vars = NULL, args = list())
}
\arguments{
\item{dist}{\code{SIMPle.dist} object}

\item{selection}{integer vector, describes new variable order/selection}

\item{vars}{number of variables to select after ordering}

\item{args}{list, named arguments passed to order_dist}
}
\value{
A \code{SIMPle.dist} object, potentially of smaller dimension
}
\description{
Select and reorder variables of a distribution
}
