% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dist.R
\name{split_data}
\alias{split_data}
\title{Split up data matrix by subgroups}
\usage{
split_data(data, by = rep(1, nrow(data)), names = unique(by), warn = 10)
}
\arguments{
\item{data}{matrix, contains (binary 0/1) data}

\item{by}{integer vector}

\item{names}{character, names of subgroups}

\item{warn}{integer, a warning is given if \code{length(unique(by)) > warn}}
}
\value{
a list of data matrices
}
\description{
Split up data matrix by subgroups
}
\details{
useful to pass data to \code{\link{update_dist}} if \code{groups(dist) > 1}
}
