% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update a prior distribution through data}
\usage{
update(dist, data, control = NULL)
}
\arguments{
\item{dist}{SIMPle.Dist object,}

\item{data}{data matrix of suitable type}

\item{control}{character, specifies adjustment method}
}
\value{
A SIMPle.Dist object, specifying the posterior distribution
}
\description{
Update a prior distribution through data
}
\examples{
+++ TODO +++
}
