% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dist.R
\name{update_dist}
\alias{update_dist}
\title{Update a prior distribution with observed data}
\usage{
update_dist(dist, data, control = list())
}
\arguments{
\item{dist}{SIMPle.Dist object,}

\item{data}{data matrix of suitable type}

\item{control}{character, specifies adjustment method}
}
\value{
A SIMPle.Dist object, specifying the posterior distribution
}
\description{
Update a prior distribution with observed data
}
