% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize a distribution or sample}
\usage{
visualize(x, group = 1, inference = NULL, subset = NULL, ...)
}
\arguments{
\item{x}{a SIMPle.dist or SIMPle.sample object}

\item{group}{integer, selects group to be plotted}

\item{inference}{SIMPle.inference object, or list of arguments passed to \code{\link{infer}}.}

\item{subset}{integer, set number of points to draw. NULL (default), draws all points.}

\item{...}{named arguments passed to \code{\link{draw_sample}} and \code{\link{GGally::ggpairs}}}
}
\value{

}
\description{
Visualize a distribution or sample
}
