% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_gamma.R
\name{construct_gamma}
\alias{construct_gamma}
\title{Construct the gamma parameter for the multivariate Beta (mbeta) distribution.}
\usage{
construct_gamma(
  vars,
  nu = 1,
  moments = nu * (matrix(1/4, vars, vars) + diag(rep(1/4, vars))),
  corr = NULL,
  tol = 0.01,
  input = "moments",
  ...
)
}
\arguments{
\item{vars}{integer, dimension of multivariate beta distribution.}

\item{nu}{numeric, equal to the sum of resulting gamma vector (prior observations).}

\item{moments}{matrix, .... The entry (j,k) of \code{cp} specifies the
expected joint probability P(B_j = B_k = 1). See details.}

\item{corr}{matrix, correlation matrix, only needs to be supplied if input="corr" (default: NULL)}

\item{tol}{numeric, if obtained momemnt matrix deviates more than \code{tol} from the target}

\item{input}{character, either "moments" (default) or "corr"}

\item{...}{a warning is given. See details.}
}
\value{
A numeric vector gamma of length 2^dims suitable as a parameter for
the multivariate Beta (mbeta) distribution.
}
\description{
Construct the gamma parameter for the multivariate Beta (mbeta) distribution.
}
\examples{
construct_gamma(3, 1)

}
