% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample_pred.R
\name{draw_sample_pred}
\alias{draw_sample_pred}
\title{Draw a sample from the predictive (mbeta-mbin) distribution}
\usage{
draw_sample_pred(
  dist,
  sample = 100,
  size = rep(1, groups(dist)),
  method_pred = c("mbeta_approx", "rmvbin"),
  method_ext = c("basic", "cov"),
  regu = 1,
  count = TRUE,
  correct = TRUE,
  msg = TRUE
)
}
\arguments{
\item{dist}{\code{SIMPle.dist} object, see \code{\link{define_dist}}}

\item{sample}{either a precomputed \code{SIMPle.sample} object or an integer specifiying the sample size}

\item{size}{sample size(s) for predictive count data}

\item{method_pred}{character, for drawing predictive sample, only "rmvbin" available}

\item{method_ext}{character, method for sample 'extesnion', only "basic" available}

\item{regu}{integer, regu parameter}

\item{count}{logical, should outcome be counts (TRUE) or probabilities (FALSE)}

\item{correct}{logical, should outcome be enforced to correspond to correct counts?}

\item{msg}{}
}
\description{
Draw a sample from the predictive (mbeta-mbin) distribution
}
\details{
Predictive mbeta-mbin distribution
}
