% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer}
\alias{infer}
\title{Conduct statistical inference based on (posterior) sample or distribution}
\usage{
infer(
  x,
  stats = "mean",
  prob = 0.95,
  aggr_var = "all",
  aggr_group = "&",
  method = c("auto", "sample", "copula", "approx"),
  ...
)
}
\arguments{
\item{x}{a \code{SIMPle.dist} or \code{SIMPle.sample} object}

\item{stats}{character, specifies statistics to be calculated}

\item{prob}{numeric, coverage probability of credible interval}

\item{aggr_var}{character, function name how cover events are aggregated across variables}

\item{aggr_group}{character, function name how cover events are aggregated across groups}

\item{method}{character, method to calculate credible regions}

\item{...}{}
}
\value{
\code{SIMPle.result} object
}
\description{
Conduct statistical inference based on (posterior) sample or distribution
}
