% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lAddIntr.R
\name{lAddIntr}
\alias{lAddIntr}
\alias{new_lAddIntr}
\alias{validate_lAddIntr}
\title{Addition of ISAs into ongoing platform trial}
\usage{
new_lAddIntr(
  fnAddIntr = function(lPltfDsgn, lPltfTrial, lAddArgs) { },
  lAddArgs = list()
)

validate_lAddIntr(x)

lAddIntr(nTrt)
}
\arguments{
\item{fnAddIntr}{Function which will add ISAs}

\item{lAddArgs}{Further arguments used in fnAddIntr}
}
\description{
Functions for creating, validating and simple use of class lAddIntr
}
\examples{

}
