% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lAddPats.R
\name{lAddPats}
\alias{lAddPats}
\alias{new_lAddPats}
\alias{validate_lAddPats}
\alias{summary.lAddPats}
\title{Add simulated Pats to ISAs}
\usage{
new_lAddPats(fnAddPats = function(lPltfTrial, lAddArgs) {
}, lAddArgs = list())

validate_lAddPats(x)

lAddPats()

\method{summary}{lAddPats}(x, ...)
}
\description{
Functions and rules for adding Pats to ISAs of class lAddPats
}
\examples{

}
