% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lAllocArm.R
\name{lAllocArm}
\alias{lAllocArm}
\alias{new_lAllocArm}
\alias{validate_lAllocArm}
\title{Within ISA allocation ratio}
\usage{
new_lAllocArm(
  fnAllocArm = function(lPltfTrial, lAddArgs) { },
  lAddArgs = list()
)

validate_lAllocArm(x)

lAllocArm()
}
\description{
Functions and rules for within ISA allocation ratio of class lAllocArm
}
\examples{

}
