% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lAllocIntr.R
\name{lAllocIntr}
\alias{lAllocIntr}
\alias{new_lAllocIntr}
\alias{validate_lAllocIntr}
\alias{summary.lAllocIntr}
\title{Between ISA allocation ratio}
\usage{
new_lAllocIntr(
  fnAllocIntr = function(lPltfTrial, lAddArgs) { },
  lAddArgs = list()
)

validate_lAllocIntr(x)

lAllocIntr()

\method{summary}{lAllocIntr}(x, ...)
}
\description{
Functions and rules for between ISA allocation ratio of class lAllocIntr
}
\examples{

}
