% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lCheckAnlsMstn.R
\name{lCheckAnlsMstn}
\alias{lCheckAnlsMstn}
\alias{new_lCheckAnlsMstn}
\alias{validate_lCheckAnlsMstn}
\title{Check Analysis Milestone}
\usage{
new_lCheckAnlsMstn(
  fnCheckAnlsMstn = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lCheckAnlsMstn(x)

lCheckAnlsMstn(bInclude = TRUE, vInfTimes = c(1), column = "OutObsTime")
}
\description{
Functions and rules for checking analysis milestone of class lCheckAnlsMstn
}
