% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lCheckAnlsMstn.R
\name{lCheckAnlsMstn}
\alias{lCheckAnlsMstn}
\alias{new_lCheckAnlsMstn}
\alias{validate_lCheckAnlsMstn}
\title{Check Analysis Milestones}
\usage{
new_lCheckAnlsMstn(
  fnCheckAnlsMstn = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lCheckAnlsMstn(x)

lCheckAnlsMstn(bInclude = TRUE, vInfTimes = c(1), column = "OutObsTime")
}
\arguments{
\item{fnCheckAnlsMstn}{Function that checks if analysis milestones are reached}

\item{lAddArgs}{Further arguments used in fnCheckAnlsMstn}
}
\description{
Functions and rules for checking analysis milestones of class lCheckAnlsMstn
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lCheckAnlsMstn()
}

}
