% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lCheckEnrl.R
\name{lCheckEnrl}
\alias{lCheckEnrl}
\alias{new_lCheckEnrl}
\alias{validate_lCheckEnrl}
\title{Check ISA Enrllment}
\usage{
new_lCheckEnrl(
  fnCheckEnrl = function(lPltfTrial, lAddArgs) { },
  lAddArgs = list()
)

validate_lCheckEnrl(x)

lCheckEnrl()
}
\description{
Functions and rules for within ISA checking of Enrllment of class lCheckEnrl
}
\examples{

}
