% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lFnDef.R
\name{lFnDef}
\alias{lFnDef}
\alias{new_lFnDef}
\alias{validate_lFnDef}
\title{List of Function Definitions}
\usage{
new_lFnDef(
  fnInitialize = function(lPltfDsgn) {
 },
  fnAddNewIntr = function(lPltfDsgn, lPltfTrial) {
 },
  fnUpdateAlloc = function(lPltfDsgn, lPltfTrial) {
 },
  fnGenNewPatData = function(lPltfDsgn, lPltfTrial) {
 },
  fnIntrAction = function(lPltfDsgn, lPltfTrial) {
 },
  fnCheckTrialClose = function(lPltfDsgn, lPltfTrial) {
 },
  fnWrapup = function(lPltfDsgn, lPltfTrial) {
 }
)

validate_lFnDef(x)

lFnDef()
}
\description{
Different ISA Designs to draw from of class lIntrDsgnList
}
