% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lFnDef.R
\name{lFnDef}
\alias{lFnDef}
\alias{new_lFnDef}
\alias{validate_lFnDef}
\title{List of Function Definitions}
\usage{
new_lFnDef(
  fnInitialize = function(lPltfDsgn) {
 },
  fnAddNewIntr = function(lPltfDsgn, lPltfTrial) {
 },
  fnUpdateAlloc = function(lPltfDsgn, lPltfTrial) {
 },
  fnGenNewPatData = function(lPltfDsgn, lPltfTrial) {
 },
  fnIntrAction = function(lPltfDsgn, lPltfTrial) {
 },
  fnCheckTrialClose = function(lPltfDsgn, lPltfTrial) {
 },
  fnWrapup = function(lPltfDsgn, lPltfTrial) {
 }
)

validate_lFnDef(x)

lFnDef()
}
\arguments{
\item{fnInitialize}{Function for initialization}

\item{fnAddNewIntr}{Function for adding new interventions}

\item{fnUpdateAlloc}{Function for updating allocation}

\item{fnGenNewPatData}{Function for generating new patient data}

\item{fnIntrAction}{Function for intervention actions}

\item{fnCheckTrialClose}{Function for checking if trial should close}

\item{fnWrapup}{Function for wrapping up the trial}

\item{x}{An object of class lFnDef}
}
\description{
Different ISA Designs to draw from of class lIntrDsgnList
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lFnDef()
}

}
