% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lInitIntr.R
\name{lInitIntr}
\alias{lInitIntr}
\alias{new_lInitIntr}
\alias{validate_lInitIntr}
\title{Initialization of ISAs}
\usage{
new_lInitIntr(
  fnInitIntr = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lInitIntr(x)

lInitIntr(
  cIntrName = "I_NoName",
  cArmNames = c("ArmNoName"),
  nMaxNIntr = 100,
  vRandList = as.vector(replicate(n = ceiling(nMaxNIntr/length(cArmNames)), expr =
    sample(cArmNames))),
  vMaxNArms = rep(ceiling(nMaxNIntr/length(cArmNames)), length(cArmNames))
)
}
\arguments{
\item{fnInitIntr}{Function which will initialize ISAs}

\item{lAddArgs}{Further arguments used in fnInitIntr}
}
\description{
Functions for creating, validating and simple use of class lInitIntr
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lInitIntr()
}

}
