% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lIntrDsgn.R
\name{lIntrDsgn}
\alias{lIntrDsgn}
\alias{new_lIntrDsgn}
\alias{validate_lIntrDsgn}
\title{ISA Design List}
\usage{
new_lIntrDsgn(
  lInitIntr = list(),
  lAllocArm = list(),
  lPatOutcome = list(),
  lCheckAnlsMstn = list(),
  lAnls = list(),
  lSynthRes = list(),
  lCheckEnrl = list(),
  ...
)

validate_lIntrDsgn(x)

lIntrDsgn(
  lInitIntr,
  lAllocArm,
  lPatOutcome,
  lCheckAnlsMstn,
  lAnls,
  lSynthRes,
  lCheckEnrl,
  ...
)
}
\description{
Different ISA Designs to draw from of class lIntrDsgnList
}
