% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lIntrDsgn.R
\name{lIntrDsgn}
\alias{lIntrDsgn}
\alias{new_lIntrDsgn}
\alias{validate_lIntrDsgn}
\title{ISA Design List}
\usage{
new_lIntrDsgn(
  lInitIntr = list(),
  lAllocArm = list(),
  lPatOutcome = list(),
  lCheckAnlsMstn = list(),
  lAnls = list(),
  lSynthRes = list(),
  lCheckEnrl = list(),
  ...
)

validate_lIntrDsgn(x)

lIntrDsgn(
  lInitIntr,
  lAllocArm,
  lPatOutcome,
  lCheckAnlsMstn,
  lAnls,
  lSynthRes,
  lCheckEnrl,
  ...
)
}
\arguments{
\item{lInitIntr}{Module for intervention initialization}

\item{lAllocArm}{Module for within-ISA allocation}

\item{lPatOutcome}{Module for patient outcomes}

\item{lCheckAnlsMstn}{Module for checking analysis milestones}

\item{lAnls}{Module for analysis}

\item{lSynthRes}{Module for synthesizing results}

\item{lCheckEnrl}{Module for checking enrollment}

\item{...}{Additional ISA-level modules}

\item{x}{An object of class lIntrDsgn}
}
\description{
Different ISA Designs to draw from of class lIntrDsgnList
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lIntrDsgn()
}

}
