% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lOCSynth.R
\name{lOCSynth}
\alias{lOCSynth}
\alias{new_lOCSynth}
\alias{validate_lOCSynth}
\title{Synthesize individual trial simulations into OCs}
\usage{
new_lOCSynth(
  fnOCSynth = function(lIndTrials, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lOCSynth(x)

lOCSynth()
}
\description{
Functions and rules for synthesizing results of class lOCSynth
}
