% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lPatOutcome.R
\name{lPatOutcome}
\alias{lPatOutcome}
\alias{new_lPatOutcome}
\alias{validate_lPatOutcome}
\title{Simulate patient outcomes}
\usage{
new_lPatOutcome(
  fnPatOutcome = function(lPltfTrial, lAddArgs) { },
  lAddArgs = list()
)

validate_lPatOutcome(x)

lPatOutcome(cGroups, dTheta, dSigma = NULL, dTrend, nLag)
}
\description{
Functions and rules for simulation of patient outcomes of class lPatOutcome
}
\examples{

}
