% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lPltfDsgn.R
\name{lPltfDsgn}
\alias{lPltfDsgn}
\alias{new_lPltfDsgn}
\alias{validate_lPltfDsgn}
\title{Platform Trial Design List}
\usage{
new_lPltfDsgn(lFnDef = list(), lIntrDsgn = lIntrDsgn, ...)

validate_lPltfDsgn(x)

lPltfDsgn(lIntrDsgn, ...)
}
\description{
Platform Trial Design List of class lPltfDsgn
}
