% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lPltfDsgn.R
\name{lPltfDsgn}
\alias{lPltfDsgn}
\alias{new_lPltfDsgn}
\alias{validate_lPltfDsgn}
\title{Platform Trial Design List}
\usage{
new_lPltfDsgn(lFnDef = list(), lIntrDsgn = lIntrDsgn, ...)

validate_lPltfDsgn(x)

lPltfDsgn(lIntrDsgn, ...)
}
\arguments{
\item{lFnDef}{Module containing function definitions}

\item{lIntrDsgn}{List of ISA designs}

\item{...}{Additional platform-level modules}

\item{x}{An object of class lPltfDsgn}
}
\description{
Platform Trial Design List of class lPltfDsgn
}
\examples{
\dontrun{
# Internal module used by platform trial simulation
x <- lPltfDsgn()
}

}
