% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lRecrPars.R
\name{lRecrPars}
\alias{lRecrPars}
\alias{new_lRecrPars}
\alias{validate_lRecrPars}
\title{Recruitment Parameters}
\usage{
new_lRecrPars(
  fRecrProc = function(dCurrTime, dActvIntr, lArgs) { },
  lArgs = list()
)

validate_lRecrPars(x)

lRecrPars(lambda)
}
\arguments{
\item{lArgs}{Arguments}

\item{fRecrPars}{Function}
}
\description{
Functions for creating, validating and simple use of class lRecrPars
}
\examples{

x <- lRecrPars(4)
validate_lRecrPars(x)
y <- list(
  fRecrProc = function(x) {x},
  lArgs = list()
)
validate_lRecrPars(y)

}
