% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lRecrPars.R
\name{lRecrPars}
\alias{lRecrPars}
\alias{new_lRecrPars}
\alias{validate_lRecrPars}
\alias{plot.lRecrPars}
\alias{summary.lRecrPars}
\title{Recruitment Parameters}
\usage{
new_lRecrPars(
  fnRecrProc = function(lPltfTrial, lAddArgs) { },
  lAddArgs = list()
)

validate_lRecrPars(x)

lRecrPars(nPat)

\method{plot}{lRecrPars}(x, dCurrTime = 1:52, ...)

\method{summary}{lRecrPars}(x, ...)
}
\arguments{
\item{lAddArgs}{Further arguments used in fnRecrPars}

\item{fnRecrPars}{Function which will simulate number of new patient arrivals in time step
dependent on the the current global variables}
}
\description{
Functions for creating, validating and simple use of class lRecrPars
}
\examples{

x <- lRecrPars(4)
validate_lRecrPars(x)
plot(x)
summary(x)

}
