% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lSimBase.R
\name{lSimBase}
\alias{lSimBase}
\alias{new_lSimBase}
\alias{validate_lSimBase}
\title{Simulation of patient Baseline data}
\usage{
new_lSimBase(fnSimBase = function(lPltfTrial, lAddArgs) {
}, lAddArgs = list())

validate_lSimBase(x)

lSimBase(
  vars = list(list("age", "max(rnorm(n, 45, 10), 18)"), list("sex",
    "sample(0:1, n, replace = TRUE)"), list("time", "t")),
  names = c("Age", "Sex", "InclusionTime")
)
}
\description{
Functions and rules for simulating patient Baseline data of class lSimBase
}
\examples{

}
