% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lStopRule.R
\name{lStopRule}
\alias{lStopRule}
\alias{new_lStopRule}
\alias{validate_lStopRule}
\alias{summary.lStopRule}
\title{Check if Platform Stopping Rules are reached}
\usage{
new_lStopRule(
  fnStopRule = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lStopRule(x)

lStopRule(nWeeks = NULL, bNoActive = NULL)

\method{summary}{lStopRule}(x, ...)
}
\arguments{
\item{fnStopRule}{Function which will check if Stop Rules are reached}

\item{lAddArgs}{Further arguments used in fnStopRule}
}
\description{
Functions for creating, validating and simple use of class lStopRule
}
