% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lSynthRes.R
\name{lSynthRes}
\alias{lSynthRes}
\alias{new_lSynthRes}
\alias{validate_lSynthRes}
\title{Synthesize trial results into decisions}
\usage{
new_lSynthRes(
  fnSynthRes = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lSynthRes(x)

lSynthRes(alpha = 0.05)
}
\description{
Functions and rules for synthesizing results of class lSynthRes
}
