% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lUpdIntrAlloc.R
\name{lUpdIntrAlloc}
\alias{lUpdIntrAlloc}
\alias{new_lUpdIntrAlloc}
\alias{validate_lUpdIntrAlloc}
\title{Update between ISA allocation ratio}
\usage{
new_lUpdIntrAlloc(
  fnUpdIntrAlloc = function(lPltfTrial, lAddArgs) {
 },
  lAddArgs = list()
)

validate_lUpdIntrAlloc(x)

lUpdIntrAlloc()
}
\description{
Functions and rules for between ISA allocation ratio update of class lUpdIntrAlloc
}
