% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_dist.R
\name{order_dist}
\alias{order_dist}
\title{Order a SIMPle.dist (!experimental!)}
\usage{
order_dist(dist, method = "prob", cpe = "min", cu = 0.75,
  lower.tail = FALSE)
}
\arguments{
\item{dist}{SIMPle.dist object}

\item{method}{"mean" or "prob"}

\item{cpe}{"min" or "prod"}

\item{cu}{numeric, cutoff for method "prob"}

\item{lower.tail}{logical, default FALSE}
}
\value{
ordering
}
\description{
Order a SIMPle.dist (!experimental!)
}
