% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize a probability distribution.}
\usage{
visualize(dist, details = 0, ...)
}
\arguments{
\item{dist}{\code{SIMPle.dist}, a distribution to visualize.}

\item{details}{integer, specfies what should be plotted. Possible values are 0 (marginal and
bivariate densities, default), 1 (quantiles CDF), 2 (extremes histogram), 3 (exceedence frequency)}

\item{...}{}
}
\value{
NONE
}
\description{
Visualize a probability distribution.
}
\examples{
+++ TODO +++
}
