% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize a distribution or sample}
\usage{
visualize(x, group = 1, inference = NULL, subset = NULL, ...)
}
\arguments{
\item{x}{a  \code{SIMPle.dist} or \code{SIMPle.sample} object}

\item{group}{integer, selects group to be plotted}

\item{inference}{\code{SIMPle.result} object (currently ignored!)}

\item{subset}{integer, set number of points to draw. NULL (default), draws all points.}

\item{...}{named arguments passed to \code{\link{draw_sample}} and \code{GGally::ggpairs},
e.g. set alpha via \code{mapping=ggplot2::aes(alpha=0.5)}}
}
\value{
GGally::ggpairs plot applied to (prior/posterior) sample
}
\description{
Visualize a distribution or sample
}
